requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
supabase==2.9.1
razorpay==1.4.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.0
python-multipart==0.0.9